+++ client/src/lib/nav-catagories.js
@@ -31,6 +31,7 @@ const {
   classicInfraTf,
   classicSecurityGroupTf,
   classicVsiTf,
+  classicBareMetalTf,
 } = require("./json-to-iac");
 const { cisTf } = require("./json-to-iac/cis");
 const { classicGatewayTf } = require("./json-to-iac/classic-gateway");
@@ -308,6 +309,15 @@ const navCatagories = [
         toTf: (config) => {
           return classicVsiTf(config) || "";
         },
+      },
+      {
+        title: "Classic Bare Metal",
+        path: "/form/classicBareMetal",
+        react_icon: "IbmCloudBareMetalServer",
+        toTf: (config) => {
+          return classicBareMetalTf(config) || "";
+        },
+        jsonField: "classic_bare_metal",
       },
     ],
   },