+++ client/src/lib/state/state.js
@@ -52,6 +52,7 @@ const { initCisGlbStore } = require("./cis-glb.js");
 const { initFortigateStore } = require("./fortigate.js");
 const { initClassicSecurityGroups } = require("./classic-security-groups.js");
 const { initClassicVsi } = require("./classic-vsi.js");
+const { initClassicBareMetalStore } = require("./classic-bare-metal.js");
 
 /**
  * get state for craig
@@ -170,6 +171,7 @@ const state = function (legacy) {
   initFortigateStore(store);
   initClassicSecurityGroups(store);
   initClassicVsi(store);
+  initClassicBareMetalStore(store);
 
   /**
    * hard set config dot json in state store