+++ CHANGELOG.md
@@ -18,6 +18,7 @@ All notable changes to this project will be documented in this file.
 - Power VS High Availability is now supported for `mad02`, `mad04`, `us-east`, `wdc06`, `us-south`, `eu-de-1`, and `eu-de-2`
 - When updating a Power VS Instance name, references to that instance are now updated to match the new name
 - Users can now upload JSON directly to CRAIG from the local file explorer from the Projects page by clicking the new `Upload JSON` button
+- Users can now create Classic Bare Metal Servers from the form page `/form/classicBareMetal`
 
 ### Fixes